moved functions to App, getting typing bug for key
